+++ Classic/module.php
@@ -3,7 +3,7 @@
 return [
     'name'        => 'Классическая главная',
     'description' => 'Классический вид главной страницы вместо ленты',
-    'version'     => '1.0.1',
+    'version'     => '1.0.2',
     'requires'    => '14.0.0',
     'author'      => 'Vantuz',
     'email'       => 'admin@visavi.net',